move shared utilities to lib
