+++ packages/fake-api/src/vite.ts
@@ -1,58 +1,17 @@
-import { URLPattern } from 'urlpattern-polyfill'
 
+
+import { Router } from './lib/utils.js'
+import type { Dependencies, FS } from './lib/utils.js'
 import type { Plugin } from 'vite'
 
-export type RestRequest = {
-  method: string
-  params: Record<string, string | readonly string[]>
-  body: Record<string, any>
-  url: {
-    searchParams: URLSearchParams
-  }
+type PluginOptions<TDependencies extends object = {}> = {
+  fs: FS<TDependencies>
+  dependencies: Dependencies<TDependencies>
 }
-
-export type MockResponse = {
-  headers?: Record<string, string>
-  body: string | Record<string, unknown>
-} | undefined
-export type MockResponder = (req: RestRequest) => MockResponse
-
-export type Dependencies<TDependencies extends object = {}> = {
-  env: <T extends string>(key: T, d?: string) => string
-} & TDependencies
-export type MockEndpoint<TDependencies extends object = {}> = <TArgs extends Dependencies<TDependencies>>(args: TArgs) => MockResponder
-export type FS<TDependencies extends object = {}> = Record<string, MockEndpoint<TDependencies>>
-
-class Router<T> {
-  routes: Map<URLPattern, T> = new Map()
-  constructor(routes: Record<string, T>) {
-    Object.entries(routes).forEach(([key, value]) => {
-      if (key.includes('://')) {
-        return
-      }
-      this.routes.set(new URLPattern({
-        pathname: key.replace('+', '\\+'),
-      }), value)
-    })
-  }
-
-  match(path: string) {
-    for (const [pattern, route] of this.routes) {
-      const _url = `data:${path}`
-      if (pattern.test(_url)) {
-        const args = pattern.exec(_url)
-        const params = args?.pathname.groups || {}
-        return {
-          route,
-          params: Object.entries(params).reduce((prev: Record<string, string>, [key, value]) => {
-            prev[key] = value || ''
-            return prev
-          }, {}),
-        }
-      }
-    }
-    throw new Error(`Matching route for '${path}' not found`)
-  }
+type FetchOptions = {
+  method?: string
+  body?: Record<string, string>
+  headers?: Record<string, string | string[] | undefined>
 }
 
 const strToEnv = (str: string): [string, string][] => {
@@ -63,16 +22,6 @@ const strToEnv = (str: string): [string, string][] => {
       const [key, ...value] = item.split('=')
       return [key, value.join('=')] as [string, string]
     })
-}
-
-type PluginOptions<TDependencies extends object = {}> = {
-  fs: FS<TDependencies>
-  dependencies: Dependencies<TDependencies>
-}
-type FetchOptions = {
-  method?: string
-  body?: Record<string, string>
-  headers?: Record<string, string | string[] | undefined>
 }
 
 export default <TDependencies extends object = {}>(opts: PluginOptions<TDependencies>): Plugin => {
